+++ docker-compose.yml
@@ -32,6 +32,7 @@ services:
       - '${SERVER_PORT}:${SERVER_PORT}'
     environment:
       SERVER_PORT: ${SERVER_PORT}
+      POSTGRES_HOST: ${POSTGRES_HOST}
 
 
   postgres:
