+++ packages/server/orm/controllers/CommentController/index.ts
@@ -1,0 +1,1 @@
+export * from './CommentController'
