+++ packages/server/orm/controllers/TopicController/index.ts
@@ -1,0 +1,1 @@
+export * from './TopicController'
