--- packages/server/orm/init/sequelizeDB/index.ts
@@ -1,1 +1,0 @@
-export * from './connectDB'
