+++ packages/server/orm/models/Reply/ReplyModel.ts
@@ -1,7 +1,7 @@
 import { Model } from 'sequelize'
 import { DataType } from 'sequelize-typescript'
 
-import { sequelize } from '../../init'
+import { sequelize } from '../../sequelizeInit'
 
 export class ReplyModel extends Model {
   public comment_id!: number
