+++ packages/server/orm/models/Topic/TopicModel.ts
@@ -1,7 +1,7 @@
 import { Model } from 'sequelize'
 import { DataType } from 'sequelize-typescript'
 
-import { sequelize } from '../../init'
+import { sequelize } from '../../sequelizeInit'
 export class TopicModel extends Model {
   public title!: string
   public replies_count?: number
