+++ packages/server/orm/routes/commentRoutes/index.ts
@@ -1,0 +1,1 @@
+export * from './commentRoutes'
