+++ packages/server/orm/routes/index.ts
@@ -1,0 +1,3 @@
+export * from './commentRoutes'
+export * from './replyRoutes'
+export * from './topicRoutes'
