+++ packages/server/orm/routes/topicRoutes/index.ts
@@ -1,0 +1,1 @@
+export * from './topicRoutes'
