+++ packages/server/orm/sequelizeInit/sequelizeDB/index.ts
@@ -1,0 +1,1 @@
+export * from './sequelizeDb'
