+++ packages/server/orm/sequelizeInit/sequelizeDB/sequelizeDb.ts
@@ -1,16 +1,21 @@
 import { Sequelize } from 'sequelize'
 import type { SequelizeOptions } from 'sequelize-typescript'
 
-const { POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB, POSTGRES_PORT } =
-  process.env
+const {
+  POSTGRES_USER,
+  POSTGRES_PASSWORD,
+  POSTGRES_DB,
+  POSTGRES_PORT,
+  POSTGRES_HOST,
+} = process.env
 
 const sequelizeOptions: SequelizeOptions = {
-  host: 'localhost',
+  host: POSTGRES_HOST,
   port: Number(POSTGRES_PORT),
   username: POSTGRES_USER,
   password: POSTGRES_PASSWORD,
   database: POSTGRES_DB,
-  dialect: 'postgres', // 'mysql', 'sqlite', 'mariadb', 'mssql'
+  dialect: 'postgres',
 }
 
 export const sequelize = new Sequelize(sequelizeOptions)
